% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatClust.r
\name{plotHeatClust}
\alias{plotHeatClust}
\title{Heat map with hierarchical clustering}
\usage{
plotHeatClust(input.path = NULL)
}
\arguments{
\item{input.path}{Path to a tab-separated txt file (note: path needs to be in
Linux style, i.e. "/" instead of "\").}
}
\value{
The desired heat map with hierarchical clustering exported
to a pdf file.
}
\description{
Plots a heat map including dendrograms generated by hierarchical clustering
for the data given via the argument "input.path" and writes it to a pdf file
in the same location.
}
\examples{
plotHeatClust("path/to/your/data-file.txt")
}
