% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.r
\name{plotPCA}
\alias{plotPCA}
\title{PCA plot}
\usage{
plotPCA(input.path = NULL)
}
\arguments{
\item{input.path}{Path to a tab-separated txt file (note: path needs to be in
Linux style, i.e. "/" instead of "\").}
}
\value{
Plot with the desired PCA plot exported to a pdf file.
}
\description{
Plots a PCA for the data given via the argument "input.path" and writes it to
a pdf file in the same location.
}
\examples{
plotPCA("path/to/your/data-file.txt");
}
